Create Word Counter
The program starts with an infinite while loop, which continues until the user chooses to quit.

The user is prompted with the question "Do you want the word counter to start (yes/no): ". They can respond with "yes" or "no" (case-insensitive).

If the user responds with "yes," they are prompted to enter a text with spaces between words.

The program splits the input text into words using whitespace as the delimiter and counts the number of words.

The word count is then printed to the console.

If the user responds with "no," the program exits the loop and prints a quitting message.

If the user provides an invalid choice (neither "yes" nor "no"), they are prompted to try again.